feat(autocomplete-controller): change the use of `config.action` and utilize UrlManager for URL
Previously `config.action` would alter the existing form action - but this prevented complete
override of a form entirely. It makes more sense to ignore the form entirely if `config.action` is
provided as the form action can easily be modified separately. Additionally the URL generated by the
`enterKey` handler was not utilizing the UrlManager, this is necessary in order to handle special
URLs that may not be using query parameters. In order to override form submission the `enterKey`
handler was moved to `keydown` instead of `keyup` in order to catch the event and prevent it. It was
also decided to move `escKey` to `keydown` as well.